Adding support for exception expectations (#7)
Adding support for the `thrown` keyword. This keyword will allow writing features that expect an exception to be thrown and to assert different things about the thrown exception. The capability already exists in unittest/pytest but this has nicer syntax. The thrown exception doesn't necessarily have to be of the exact same time as the expected one. It could also be a subtype